Drive the last deadline tests from the same clock
#115 fixed the insertion family. The two tests that failed on its own CI
run were in other files: TaskWaitingTests and PromptShapingServiceTests
each raced a Task.sleep against real time and asserted the elapsed time
was small, which asserts the runner was fast enough.

InsertionClock is generalized to DeadlineClock and taken by awaitValue
and PromptShapingService as well, because all three are the same
mechanism racing a sleep and a second near-identical seam would only
mean a second fake to keep in step. DrivenClock moves out of
TextInsertionServiceTests for the same reason.

Five tests drop from wall-clock bounds of two and five seconds, and from
150ms of real sleeping, to under 100ms each. Verified by mutation: the
timeout paths were broken three ways and the tests caught all three. The
first rewrite of returnsAsSoonAsTheTaskFinishes did not catch one of
them, passing on a race instead, so it now reports whether the task had
finished at the moment the wait returned.

Closes #116
